flytt database funksjon
